% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsga3.R
\docType{package}
\name{nsga3-package}
\alias{nsga3}
\alias{nsga3-package}
\title{NSGA III for Multi Objective Feature Selection}
\description{
An adaptation of Non-dominated Sorting Genetic Algorithm III for multi objective 
    feature selection tasks. Non-dominated Sorting Genetic Algorithm III is a genetic algorithm
    that solves multiple optimization problems simultaneously by applying a non-dominated sorting 
    technique. It uses a reference points based selection operator to explore 
    solution space and preserve diversity. See the original paper by K. Deb and 
    H. Jain (2014) <DOI:10.1109/TEVC.2013.2281534> for a detailed description.
}
\references{
K. Deb, H. Jain (2014) <DOI:10.1109/TEVC.2013.2281534>
}
\author{
Artem Shramko <art.shramko at gmail.com>
}
\keyword{algorithm,}
\keyword{feature}
\keyword{genetic}
\keyword{internal}
\keyword{multi}
\keyword{nsga-iii}
\keyword{nsga3,}
\keyword{objective}
\keyword{omptimisation,}
\keyword{optimisation,}
\keyword{preprocessing,}
\keyword{selection,}
\keyword{wrapper,}
