% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsga3.R
\docType{package}
\name{nsga3-package}
\alias{nsga3}
\alias{nsga3-package}
\title{nsga3: An Implementation of Non-Dominated Sorting Genetic Algorithm III for Feature Selection}
\description{
This package provides an adaptation of NSGA III for multi objective 
    feature selection tasks. NSGA III is a genetic algorithm that solves multiple 
    optimisation problems simultaneously by applying a non-dominated sorting 
    technique. It uses a reference points based selection operator to explore 
    solution space and preserve diversity. See the original paper by K. Deb and 
    H. Jain (2013) for a detailed description.
}
\keyword{internal}
