% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{nsga3fs}
\alias{nsga3fs}
\title{NSGA III for Multi-Objective Feature Selection}
\usage{
nsga3fs(df, target, obj_list, obj_names, pareto, pop_size, max_gen, model,
  resampling, num_features = TRUE, mutation_rate = 0.1,
  threshold = 0.5, feature_cost = FALSE, cpus = 1)
}
\arguments{
\item{df}{An original dataset.}

\item{target}{Name of a column (a string), which contains classification target variable.}

\item{obj_list}{A List of objective functions to be optimisied.
Must be a list of objects of type closure.}

\item{obj_names}{A Vector of the names of objective functions.
Must match the atguments passed to pareto.}

\item{pareto}{A Pareto criteria for non-dominated sorting. Should be passed in a form:
> \eqn{low(objective_1)*high(objective_2)}
See description of rPref::low for more details.}

\item{pop_size}{Size of the population.}

\item{max_gen}{Number of generations.}

\item{model}{A mlr::makeLearner object. A model to be used for classification task.}

\item{resampling}{A mlr::makeResampleDesc object.}

\item{num_features}{TRUE if algorithm should minimise number of features as one of objectives.
You must pass a respective object to pareto as well as obj_names.}

\item{mutation_rate}{Probability of switching the value of a certain gene to its opposite.
Default value 0.1.}

\item{threshold}{Threshold applied during majority vote when calculating final output.
Default  value 0.5.}

\item{feature_cost}{A vector of feacure costs. Must be equal ncol(df)-1.
You must pass a respective object to pareto as well as obj_names.}

\item{cpus}{Number of sockets to be used for parallelisation. Default value is 1.}
}
\value{
A list with the final Pareto front: selected subsets and their respective fitness values.
}
\description{
An adaptation of NSGA III for multi objective feature selection tasks.
NSGA III is a genetic algorithm that solves multiple
optimisation problems simultaneously by applying a non-dominated sorting
technique. It uses a reference points based selection operator to explore
solution space and preserve diversity. See the paper by K. Deb and
H. Jain (2013) for a detailed description of the algorithm.
}
